Add tax_benefit_system.add_variables_from_module
Fix reform test
